Add minimal Redux setup
- Install Redux, React-Redux, and their types
- Update module resolution to "node" to find files properly (can't remember the issue)
- Add libraries to webpack vendors list
- Add Redux

To add Redux, we create the index.tsx file in "data" that for now will house the only reducer. Most of the issues come from getting types correct, which is why string types are needed.
Then in test-component, we bind the state (from the store) to the props along with the dispatcher actions.